+++ src/java/org/apache/commons/pool2/impl/GenericObjectPool.java
@@ -1292,8 +1292,8 @@ public class GenericObjectPool<T> extends BaseObjectPool<T>
             Long time = iter.next();
             if (time != null) {
                 counter++;
-                result = result * ((counter - 1) / counter) +
-                        time.longValue()/counter;
+                result = result * ((counter - 1) / (double) counter) +
+                        time.longValue()/(double) counter;
             }
         }
         return (long) result;
